requirements.txt
Flask==0.12.2
Pillow==4.2.1
boto3==1.4.7
cognitive-face==1.3.1
google-cloud-vision==0.30.0